double to decimal

+++ Models/DTO/Accounts/AccBillPaymentDto.cs
@@ -9,13 +9,13 @@ namespace Models.DTO.Accounts
     {
         public int BillId { get; set; }
         public int PaymentTypeId { get; set; }
-        public double TotalAmount => (CashAmount??0) + (ChequeAmount??0);
+        public decimal TotalAmount => (CashAmount??0) + (ChequeAmount??0);
         public int? BankAccountId { get; set; }
         public string ChequeNo { get; set; }
         public DateTime PaymentDate { get; set; }
         public string Remarks { get; set; }
-        public double? CashAmount { get; set; }
-        public double? ChequeAmount { get; set; }
+        public decimal? CashAmount { get; set; }
+        public decimal? ChequeAmount { get; set; }
         public InvPurchaseMasterDto Bill { get; set; }
 
         public string PaymentTypeText =>
