+++ Models/DTO/Reporting/Sales/RptTaxCollectionDto.cs
@@ -23,6 +23,6 @@ namespace Models.DTO.Reporting.Sales
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double TaxAmount { get; set; }
+        public decimal TaxAmount { get; set; }
     }
 }
