+++ POS_API/Data/AccLedgerPosting.cs
@@ -12,9 +12,9 @@ namespace POS_API.Data
         public int TransactionDetailId { get; set; }
         public string VoucherNo { get; set; }
         public string Description { get; set; }
-        public double Dr { get; set; }
-        public double Cr { get; set; }
-        public double Balance { get; set; }
+        public decimal Dr { get; set; }
+        public decimal Cr { get; set; }
+        public decimal Balance { get; set; }
         public DateTime? PostedOn { get; set; }
         public int? PostedBy { get; set; }
