+++ POS_API/Data/InvGrrnDetails.cs
@@ -9,8 +9,8 @@ namespace POS_API.Data
         public int GrrnId { get; set; }
         public int ItemId { get; set; }
         public string BatchNo { get; set; }
-        public double ReturnQuantity { get; set; }
-        public double Rate { get; set; }
+        public decimal ReturnQuantity { get; set; }
+        public decimal Rate { get; set; }
         public int CompanyId { get; set; }
         public int Status { get; set; }
         public int? CreatedBy { get; set; }
